scheduling-app/backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.30
Flask-Migrate==4.0.7
Flask-Cors
python-dotenv
Flask-SocketIO